Autodetect TestFramework if none is configure explicitly
When the TestFramework of a Test task is not configured explicitly it
used to default to `useJUnit()`. Now, the task's classpath is inspected
for a junit-platform-engine.jar or testng.jar. If one of them is found,
the task uses the JUnit Platform or TestNG to execute tests,
respectively.

Since we have internal dependencies on the JUnit Platform and TestNG,
all Test tasks in our own build are configured to `useJUnit()` and
explicitly.