Only keep artifacts from pom if the file path is unambiguous
If the packaging indicated in a pom is not in the list of 'known
jar packagings', we assume that the artifact could have the extension
indicated by the packaging. We first test if that artifact exists
with a HEAD request, and only if it does not, we go for the 'jar'
artifact.

Since using a variant file rule disables this mechanism, we remove
such ambiguous artifacts from the modified artifact list to give users
the chance to explicitly state which artifact to expect in the rule
they add anyway.