Detect removed methods only in public superclass
There were some slight problems in the
`MethodsRemovedInInternalSuperClassRule` causing to report methods
removed from a super class in the public API in all its subclasses.

For example, when removing two deprecated methods from `DefaultTask`,
the removal was needed to be accepted in all task implementations.

Moreover, we only checked the signature of a removed method not the
name of the method.