Remove the embedded-kotlin-repository
It was based on ClientModule dependencies that caused more troubles than
anything. The main one being that the artificial dependency graph needed
to be kept up to date with each Kotlin version. Moreover since several
Kotlin versions, declaring a repository is necessary in any case. This
change simplifies the code and should make the runtime a bit faster.

However, the Kotlin dependencies when applying the `kotlin-dsl` plugin
are now required to be downloaded. This isn't a big change and should
only impact people using the `kotlin-dsl` plugin without depending on
Kotlin in other parts of their build. This could be alleviated in the
future if dependency resolution considers the Gradle install/distro as a
source of artifacts in the same way it uses maven local.

Along the way, the pinning of Kotlin dependencies to the embedded
version for the build scripts classpath has been moved from a resolution
rule to proper, faster, dependency constraints.

Signed-off-by: Paul Merlin <paul@gradle.com>