Rewrite section on component metadata rules
The section was written when the very first version of rules was
introduced and since then only marginally updated.

This is a complete rewrite of the section focusing on explaining
all the metadata modeling features of Gradle Module Metadata
which can be utilized in rules to enrich existing metadata.
The features are described on using real-world use cases.

Related sections are also updated where applicable.