Fix executing Specs tests via JUnit Platform and Vintage engine
Prior to this commit, the JUnitPlatformTestExecutionListener parsed the
unique ID to determine the class and method name for TestIdentifiers
without children but with a ClassSource. This occurs when a Spock test
class uses `@Unroll` on all its test methods. However, other testing
frameworks such as Specs2 don't provide a unique ID that is parseable
which lead to the methodName being null which cause a failure during
serialization. The effect was that tests hung because the build VM
was waiting for the worker to finish the tests it started.

The Spock-specific workaround in our listener is now replaced with a
more robust one that always uses the methodName provided by the JUnit
Platform instead of parsing it.