Fix application of dependency rules in realized metadata version
The rules for dependencies and dependency constraints were not applied
if the variant itself was added by a rule.

This commit also no longer attempts to apply any rule to the realized
version of 'pure maven configurations'. For Maven, we always use the
derived variants and always use variant matching. Therefore, rules are
never applied to the legacy configurations, but only to the derived
variants.

Follow up to #10368