Rename inheritStrictVersions() -> endorseStrictVersions()
This is more clearly communicating the semantics of the feature
from a users point of view.

The commit also removes all mentions of 'inheriting' AND 'forSubgraph'.
There have been some leftovers in documentation/comments that
will be misleading in the future. To make sure we catch all,
I also updated all variable/method/package names.