Consolidate type validation
Use ValidateContext with a few standard implementations instead of the many different interfaces and implementations we used to collect valdiation problems before. This fixes issues with types being displayed inconsistently.

The implementation is not fully correct yet: we now fail to display the correct type for problems that are bound to a type instead of a property.