Move 'composite builds' into 'Authoring Gradle Builds' chapter
Composite builds are primarily a feature to organize your own
builds into multiple independent units.
They can be used to substitute a module dependency with a local
fork. Thus the chapter is now referenced from dependency management
in the section where we describe the different types of dependencies
that are supported.