Fix duplicate classes on classpath using test fixtures
This commit works around duplicate classes found on compile
and runtime test classpath whenever the "test fixtures"
plugin is used.

This is due to the fact the Java plugin systematically adds
the main source set output to the test compile and runtime
classpath. But if we have test fixtures, they themselves
depend on the main component, so it ends up adding them
twice, possibly once as a jar and once as a class directory.

Fixes #10872