Use a custom meta tag in html to track adoc source path
This is a follow up to #10841.

Instead of using the 'description' tag, which is a standardised
tag other tools might read and display as page description,
we now use a custom name (adoc-src-path).
This also simplifies the build (and partly reverts #10841) as we
now use the 'docinfo' configuration of asciidoc to generate and include
a custom html header file for each 'adoc' file. Instead of modifying
the adoc files themselves.