Do not define scope for dependencyManagement entries
The semantics between the configuration of a constraint in Gradle and the scope
of a declaration in dependencyManagement in Maven are fundamentally
different.
Given this, Gradle will no longer attempt to define a scope for
dependencyManagement entries when creating POM files.
The only exception is the import scope as it carries special meaning.

Fixes #10878