Introduce `failOnDynamicVersion`
This commit introduces a new dependency graph validation mode,
which will make sure that if dynamic versions are found in the
graph, then either they are superceded by another version (they
don't participate in selection) or the build should fail.

This means that, for example, if a version selector uses a
version range `[1.0, 2.0[`, the build will fail because in a
subsequent build the resolution may change.

However, if there are two selectors participating, say
`[1.0, 2.0[` and `1.5`, then we choose `1.5` because this version
is within the range. Even if newer versions are released, we
would _not_ change the resolution result.