Fix snapshot handling with Gradle Module Metadata
This commit fixes a couple of bugs:

1. if Gradle Module Metadata was published and consumed, then
the `changing` flag for the resolved component metadata wouldn't
be set to `true`, which means that snapshot would effectively be
considered as persistent

2. the publish test fixtures were not using the right, timestamped,
version id for the metadata and artifacts in case of unique snapshots,
which caused the resolution to fallback to the POM file

In addition, this fixes the generated module metadata file which
was uploaded _without_ substution the the SNAPSHOT version with
the timestamped version when published on external repositories.

Finally, this highlighted a couple of issues with test fixtures
which were still using Gradle Module Metadata when they shouldn't.

Fixes #10916