Make it impossible to sign GMM file if snapshot
This commit, hopefully temporarily, makes it impossible to sign
a Gradle Module Metadata file if the version is a snapshot.

The reason is that the signature which would be produced would
be for the raw, unmodified module file. However, the final
file would be modified in case of a snapshot to include the
timestamp version instead of the `-SNAPSHOT` version for artifacts,
which means that the signature wouldn't match anymore.

To avoid this, we simply disallow signing in this case.