Add a `add-plugin` CLI option
This commit introduces a new CLI flag, `--add-plugin`, which allows adding a plugin to a build
directly from the command line. The main advantage of this is that there's no need to have a
build file to be able to download an apply a plugin.

There are different use cases for this, but mainly, this is about _bootstraping_ plugins.
For example, the vert.x team could publish a plugin which generates a templated Gradle build.
All the user would have to do would be something like:

`gradle --add-plugin com.vertx.bootstrap:1.5`

and then the plugin would take care of generating a build.

Another use case is to add diagnostics (build scans is an example of this but there's already
a built-in mechanism, --scan, to do this).

This spike is _compatible with included builds_, meaning that you can bootstrap with
a plugin currently in development using `--include-build`.