Use a linked hash set for dependency artifacts
The order can make a difference in repository selection, when
checking if an artifact-only-component containing the artifacts
exists in a repository. In absence of a metadata file, we initially test
for the existence of one artifact to decide if a repository contains
the corresponding artifact-only-component. This is always the first
artifact in the set (which is internally converted into a list).
This change makes sure that the first artifact is always the same
for a build that does not change.