Let Gradle 6.0 work with Zinc 1.3.0
The current implementation of `org.gradle.api.internal.tasks.scala.ZincScalaCompilerFactory#getCompiler` calls `ZincUtil#getDefaultBridgeModule`,
which is an internal class, but more importantly, it's been removed in Zinc 1.3.0.

There is probably a better way to prevent the use of Zinc internal classes, but moving the functionality to determine the bridge module name
feels good enough and not too intrusive given the short time left for Gradle 6.0.

A quick try to pull in zinc 1.3.0 via the `zinc` configuration works with the patch, but fails without it.

Signed-off-by: Robert Stupp <snazy@snazy.de>