Fix handling of “no settings” case WRT auto applied plugins and classpath
Some invocation args (e.g. -b) cause us to create two settings objects,
with the subsequent being empty.

Previously we were applying the auto
apply plugins to both, and inheriting the empty's classpath into the
rest of the build.

Now, we don't auto apply any plugins to the empty and have it inherit
the real settings script's classpath (which also makes it available to
the rest of the build tree).

Closes #11101