Increase timeout in integration test
Since this tests an unstable/broken Gradle process, it alway has
to spin up its own daemon. If the test runs in an environment short
on resources - for example if it runs with several other tests in
parallel - starting up the daemon can take longer. Thus we allow
the test to wait for up to 2 minutes now.

Fixes: https://github.com/gradle/build-tool-flaky-tests/issues/10