Increase timeout in integration test
This test runs several build in parallel and thus spins up multiple
daemons. If the test runs in an environment short on resources -
for example if it runs with several other tests in parallel - starting
up all the daemons can take longer. Thus we allow the test to wait for
up to 2 minutes now.

Fixes: https://github.com/gradle/build-tool-flaky-tests/issues/14