Retain stacktrace when test method is inherited
By default Gradle is trying to find "test class name" in the stacktrace (which is smart!), however,
there are cases when test class name never appears in the stacktrace.
For instance, it might happen when test class uses inheritance.
The inheritance is typically used to test over different implementations.

The fix is to retain the stacktrace in case none of the lines match.

fixes #11220

Signed-off-by: Vladimir Sitnikov <sitnikov.vladimir@gmail.com>