Normalize APP_HOME in the Windows app start script
Normalizing an app home like

    C:\app\cli\build\install\ort\bin\..

to

    C:\app\cli\build\install\ort

results in a more readable and shorter path. The latter is especially
useful to shorten the total length of the CLASSPATH variable which uses
APP_HOME to refer to all required JAR libraries. This in turn might help
to avoid running into Windows' command line length limitation when
java.exe is called with CLASSPATH.

Also see the related discussion in issue #1989.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>