Desugar producer attribute if the requesting attribute is desugared
This can be the case if a attribute on a dependency is published
and the resolved target of the dependency is a local project.
For example, a published platform dependency to a local java-platform
project.

We support 'Named' and 'Enum' for desugaring as that are the only
non-primimitve types we currently allow to be published in Gradle
Module Metadata.