Avoid eager creation of `ResolveIvyFactory`
The `ResolveIvyFactory` instance was created too eagerly. In general,
this isn't a problem because most builds will eventually resolve
dependencies. But there are cases, like when calling `help`, when this
is not the case.

The previous code was assuming that a factory would always be created,
and by side effect moved the creation of the factory to an earlier
phase during build initialization, triggering the creation of the file
access journal.

This is no longer the case: we only create the dependency verification
parameter override eagerly, and configure the build to shut it down
at the end of the build.