Fail eagerly when accessing artifacts
This commit reworks the moment when we fail because of a dependency
verification error: instead of failing at the end at the build, we
fail immediately after trying to access an artifact.

Because it's effectively impossible to do something with the artifacts
before _all_ artifacts of an artifact collection have been accessed
(because of how Gradle resolves file collections), we can collect as
much errors as possible during the resolution of a configuration and
fail immediately after, which is more user friendly than failing on
each artifact of a set.