Add lenient dependency verification mode
This mode has been added because dogfooding the feature showed
it could be painful to udpate the dependency verification
metadata file without such an option: often we want to diagnose
why a dependency is here, but as soon as the dependency
verification file is present, verification is active and immediately
fails the build.

This effectively prevents from using the usual tooling to check
where a dependency comes from. The workaround was to temporarily
rename the file, run the build, then rename again, which was tedious.

So, instead of adding a mode where verification would be totally
ignored, this commit introduces a mode where the errors are turned
into warnings. This doesn't totally silence the problems, which
makes them more visible to the developer.