Fix cycle in 'buildNeeded' and 'buildDependents' tasks
With the introduction of capabilities, it is possible that a project
depends on another variant of itself. This is the case, for example,
if you apply the 'java-test-fixtures' plugin.
The issue was that in these cases the 'buildNeeded' / 'buildDependents'
tasks would add a dependency to themselves. This is fixed by checking
that the task about to be added as dependency is not the same task for
which the dependencies are calculated.

Fixes #10410