Use file names instea of Ivy artifact names for comparison
This commit reworks the generation of verification file and
verification itself in order to use the file name instead of
the Ivy artifact name. This is done because in case of Gradle
module metadata, the file name of an artifact is not necessarily
directly bound to the module name and causes comparison issues.