Rename `BuildEventsListenerRegistry.subscribe()` to `onTaskCompletion()` to make the current contract more explicit.
At some point, we may may allow plugins to receive finish events from additional operation types, in which case new methods can be added and this method left as a convenience or deprecated.

Also add an internal `onBuildOperationComplete()` method, which gradle-profiler can use to subscribe to build operations in an instant execution friendly way. The tooling API event stream doesn't carry quite enough information for gradle-profiler at the moment. So add this method instead of keeping the work-arounds in instant execution and complexity in gradle-profiler. We may replace this method with a public one once the tooling API event stream carries the required information.