Track composition of Maven module metadata
The binary representation of a Maven POM was only tracking the
POM file itself as the source. However, a POM can reference
parent POMs which also participate in resolution. This commit
changes the `ModuleSources` so that we can query more than one
module source of a specific type, allowing us to store more
than a single module source for a single file.