WIP - add some bean properties to `Provider` and `Property` to make their use from Groovy and Kotlin more convenient.
Add read-only `value` property to `Provider<T>` to represent the current vaule of the provider. This is the same as `get()`.

Add mutable `value` property to `Property<T>` to allow the current value of the property to be set and queried. This is the same as `get()` and `set(T)`.

Add mutable `source` property to `Property<T>` to allow the current _source_ of the property to be set. This is the same as `this` and `set(Provider<T>)`.

The intention is that `get()` and `set()` would be deprecated and replaced by these new methods.