Move checksum service to build session scope
and make it use the build scope file inspector:
we can't use the global one because it's bound
to a single hasher which is supposed to be the
file hasher. At the same time our build session
scope checksum hasher will be closed _before_
the inspector invalidates entries, so we need
to make sure our caches use a different inspector.