Ignore trusted modules when writing the file
This is particularly useful when you want to ignore some modules
in the output file (for example, the native-platform snapshots)
but don't want them to appear in the verification file either.

The checking is also slightly changed so that we would only check
if a module is trusted if there's actually a verification error.