Create a different verification file if dry run is on
This commit is another usability issue: this allows simulating
the write of a verification file by resolving all configurations,
but instead of modifying the file it actually generates a different
one which can then be used for diff'ing.

This can be useful whenever the user wants to have comments and a
specific order in their verification file and want to merge the
diff by hand. It's also useful to avoid having to actually execute
the build to check if all configurations would be ok.