Split the `signing` module
This commit splits the `signing` module into another `security`
project so that we can reuse some code for dependency verification.
It's worth noting that some of the classes had to remain in the `plugins`
package because they were public APIs.

Further effort to split them out may be done later.