Do not deprecate 'JavaPlatform'
The interface is used by the Java compile task and for transferring
Java version information from the task to the compiler.
The type is now de-incubated. In the future, we might replace
(and then deprecate) or extend it.

This commit removes documentation of the interface that refers to it's
usage in the jvm-component DSL, as this usage of the interface is indeed
deprecated.