Add support for globally trusted keys
A globally trusted key can be used to trust a number of
modules and greatly simplifies configuration: instead of
having to specify checksums for all modules, a user can
declare the keys they trust and use a similar syntax to
trusted artifacts to say to what group/name/version the
key applies.

It's often the case that the same keys are used for
several artifacts of the same group or same company, so
this makes it possible to avoid a lot of boilerplate as
long as the artifacts are signed by the same keys.