Sort entries when building the verification map
This is from user feedback: it seems to be easier to read/update the
contents of the verification file if entries are sorted. If it's done
correctly, then a new version of a module would be written close to
the existing one, making it easier to do manual cleanup of the file.