Add ability to ignore keys for a specific artifact
The use case for this is whenever signature for an artifact fails, but
for some reason the user still trusts the artifact. For example, a POM
file is different between different repositories because it happened
to be published twice with different timestamps.

In this case it is recommended to ignore the signature, however we
_will_ fallback on checksum verification.