Make it possible to declare a group of repositories for exclusive content
This should address the case where an artifact can be found in a specific
group of repositories, not in a single one. This is for example the case
when an artifact can be found either in the 'release' or 'snapshot'
repository of a company, but not on the public repositories.