Restructure project tree construction so that the containing `ProjectState` for a project is always available, to avoid throwing and then silently ignoring an `IllegalArgumentException` for each project.
Also reuse more project construction logic and remove a bunch of duplicate ad hoc code. This change means that the `ProjectState` object now owns the mutable `Project` object, as was always the intention, rather than existing as a parallel tree. This can continue to evolve so that all access to the mutable project object is mediated by the `ProjectState` that owns it.