Fix circular dependencies when project have the same name
Before this commit, during dependency resolution, a synthetic
module version identifier was generated by project, using the
group and name of the project. However, it's possible for a
project in gradle to have the same name as another in the
same build, leading to duplicates. In this case the projects
were mixed together and lead to a circular dependency.

This commit fixes the problem by making sure we generate
distinct module version identifiers for such projects, by
using the full project path as the name instead of the short
name.

This also makes it possible to publish valid publications
when using the maven or ivy publish plugins. However, we detect
this problem early and warn the user that they should overwrite
the project identity in this case.