Workaround for a bug in 6.1.x
The bug (#11971) makes the build under test, called by the TAPI,
fail because the working directory is the working directory of
Gradle itself. This causes the test to try to read the verification
metadata file of Gradle itself, which exists, instead of not finding
any verification file like it should, because the build under test
doesn't have any.

The workaround is to change the working directory from the build
under test.