Properly isolate tests from the Gradle user home
When running Gradle, we use the current working directory as the
base directory for searching a `settings.gradle` file. For tooling
API tests this working directory is the one of Gradle itself, which
causes a number of issues because the root directory of the project
doesn't match the working directory.

Therefore, Gradle finds the `settings.gradle` file of its own build,
instead of finding the one in the test.

As a workaround, TAPI tests now use an action which will set the
working directory to the test directory, which should cover most
cases. Additional cases like missing settings files in projects
under tests have to be handled separately, as demonstrated in this
commit for the ProjectBuilder test.