Pass executable setting to Scala Compiler
JDK8 is the only supported java runtime to execute some versions of the
scala compiler on, which means that any gradle project using a newer
JDK version cannot compile scala code for older versions of scala
unless the executable for the scala compile task is respected.

This PR copies the executable setting from the java fork options on the
scala compile specification to the fork options used to launch the
compile daemon process.

Signed-off-by: James Baiera <james.baiera@gmail.com>