Track currently deselecting node to break another case of local cycle
If we get to a 'reselect endorsing node' situation through a node
that is deselected and from which the endorsed node (platform node)
is a child, we can not go back to that node as it is responsible
for triggering the reselection in the first place.
Without this fix, the endorsing node was attempted to be cleaned
several times causing a ConcurrentModificationException on the
outgoing edges list (without that hard break, it would have run
in an endless loop).