Add test case showing errors with Kotlin DSL
This commit mostly introduces a new (experimental) test fixture
which aims at making it possible to generate both Groovy and Kotlin
DSL scripts and therefore running integration tests with both enabled.

It's obviously not for all tests, but in this case it highlights that
the Kotlin DSL has several issues with the dependency constraint handler.

See #9249