Fix Kotlin DSL constraints block
This commit introduces a workaround for the Kotlin DSL
which doesn't delegate the constraints block to the right
delegate type. The consequence is that constraints were
mixed with dependencies when using extension methods defined
on the `DependencyConstraintHandlerScope`, because the compiler
didn't realize that the receiver type was indeed this scope.