Do not serialize multiple copies of a `CopySpec` to the instant execution cache, for a given task.
Previously, certain specs and their children would be written several times. For example, the `metaInf` spec for a `Jar` task would be written twice, which includes the manifest generation.

This not only avoids the cost of writing the same data several times but also the cost of calculating the spec contents several times, which can be expensive.