Introduce an HTML report for dependency verification
This commit introduces a dependency verification report.
This report is generated for each build which has at least
one dependency verification fatal failure. The goal is to
replace the full, plain text report with a more concise
text block, less intimidating, and redirect to the HTML
report for completeness.

Currently, this report is _always_ generated, and the
console report is left untouched. However, another commit
will replace the console report with a shorter version,
with the ability to configure to build to always use the
long console report version (in case for example it's not
simple to retrieve an HTML report from a CI build).

The HTML report explains what errors the user is facing
in context, and links to the documentation for further
details.