Change layout of HTML verification report
Use an accordion instead of the left navbar. This has a couple of
advantages:

- in most cases, there will be a single configuration so it's a
waste of space to reserve a navigation bar on the left. There
will only be multiple configurations if the report is generated
in _lenient_ mode
- it makes the report more readable on lower screen resolutions:
when reducing the width, the previous display could overlap text
and make the navbar appear beyond the main display